progress toward new trait system for SearchRaw

+++ src/Routines/LibrarySearch/methods/queryFragmentIndex.jl
@@ -198,7 +198,7 @@ function searchScan!(prec_id_to_score::Counter{UInt32, UInt8},
                     irt_high::Float32, 
                     mass_err_model::MassErrorModel,
                     quad_transmission_func::QuadTransmissionFunction,
-                    isotope_err_bounds::Tuple{Int64, Int64}
+                    isotope_err_bounds::Tuple{UInt8, UInt8}
                     ) where {U<:AbstractFloat}
     prec_min = U(getPrecMinBound(quad_transmission_func) - NEUTRON*first(isotope_err_bounds)/2)
     prec_max = U(getPrecMaxBound(quad_transmission_func) + NEUTRON*last(isotope_err_bounds)/2)
